+++ src/obsolete/aSunCalcMeanAnomaly.cpp
@@ -10,7 +10,7 @@
 //
 // TODO - describe what we're calculating here...
 //
-void aSun::aSunCalcMeanAnomaly(double T)
+void aSun::aSunCalcMeanAnomaly(long double T)
 {
 	aSunMeanAnomaly = fmod((357.52911 + T * (35999.05029 - 0.0001537 * T)), 360.);;
